chore: Update to RAML as of 7942537a940b53819589e48a59e059ff708d5698

+++ src/commercetools/platform/client/search/by_project_key_product_projections_search_request_builder.py
@@ -54,7 +54,7 @@ class ByProjectKeyProductProjectionsSearchRequestBuilder:
         self,
         *,
         fuzzy: bool = None,
-        fuzzy_level: float = None,
+        fuzzy_level: int = None,
         mark_matching_variants: bool = None,
         filter: typing.List["str"] = None,
         filter_facets: typing.List["str"] = None,
